fix: fixed export of fn
