+++ packages/sdk/src/thor/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './const';
 export * from './units';
 export * from './nonce';
-export * from './normalizeArgs';
